% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{coord_transform}
\alias{coord_transform}
\title{Transform normalized positions through a coordinate system}
\usage{
coord_transform(coord, ...)
}
\arguments{
\item{coord}{A \link{Coord} subclass instance.}

\item{...}{Method arguments; all methods take \code{x} and \code{y}, numeric
vectors of normalized positions.}
}
\value{
List with transformed \code{x} and \code{y}.
}
\description{
Transform normalized positions through a coordinate system
}
