% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{CoordCartesian}
\alias{CoordCartesian}
\title{CoordCartesian: the identity coordinate system}
\usage{
CoordCartesian(flip = FALSE)
}
\arguments{
\item{flip}{Swap x and y (see \code{\link[=coord_flip]{coord_flip()}}).}
}
\value{
An S7 object of class \code{CoordCartesian}; usually built by \code{\link[=coord_cartesian]{coord_cartesian()}} or \code{\link[=coord_flip]{coord_flip()}}.
}
\description{
CoordCartesian: the identity coordinate system
}
