% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms-base.R
\name{geom_blank}
\alias{geom_blank}
\title{Blank layer}
\usage{
geom_blank(mapping = NULL, data = NULL)
}
\arguments{
\item{mapping, data}{Standard layer overrides.}
}
\value{
A \link{Layer} to add with \code{+}.
}
\description{
Draws nothing. Useful to establish scales without showing data, or as a
placeholder in code that conditionally adds a layer.
}
\examples{
ggplot3(cars, aes(speed, dist)) + geom_blank()
}
