% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms-special.R
\name{geom_errorbar}
\alias{geom_errorbar}
\title{Error bar layer (\code{ymin} to \code{ymax} with caps)}
\usage{
geom_errorbar(
  mapping = NULL,
  data = NULL,
  color = NULL,
  alpha = NULL,
  linewidth = NULL,
  width = NULL
)
}
\arguments{
\item{mapping, data, color, alpha}{As in \code{\link[=geom_point]{geom_point()}}.}

\item{linewidth}{Stroke width.}

\item{width}{Cap width in x data/slot units.}
}
\value{
A \link{Layer}.
}
\description{
Error bar layer (\code{ymin} to \code{ymax} with caps)
}
