% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_histogram}
\alias{geom_histogram}
\title{Histogram layer}
\usage{
geom_histogram(
  mapping = NULL,
  data = NULL,
  color = NULL,
  alpha = NULL,
  bins = 30,
  binwidth = NULL,
  position = "stack"
)
}
\arguments{
\item{mapping, data, color, alpha}{As in \code{\link[=geom_point]{geom_point()}}.}

\item{bins}{Number of bins (ignored when \code{binwidth} is given).}

\item{binwidth}{Bin width in data units.}

\item{position}{\code{"stack"} (default), \code{"dodge"}, or \code{"identity"}.}
}
\value{
A \link{Layer}.
}
\description{
Histogram layer
}
