% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_segment}
\alias{geom_segment}
\title{Segment layer (\code{x},\code{y}) to (\code{xend},\code{yend})}
\usage{
geom_segment(
  mapping = NULL,
  data = NULL,
  color = NULL,
  linewidth = NULL,
  alpha = NULL
)
}
\arguments{
\item{mapping, data, color, alpha}{As in \code{\link[=geom_point]{geom_point()}}.}

\item{linewidth}{Stroke width in px.}
}
\value{
A \link{Layer}.
}
\description{
Segment layer (\code{x},\code{y}) to (\code{xend},\code{yend})
}
