% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_violin}
\alias{geom_violin}
\title{Violin layer}
\usage{
geom_violin(
  mapping = NULL,
  data = NULL,
  color = NULL,
  alpha = NULL,
  width = 0.9
)
}
\arguments{
\item{mapping, data, color, alpha}{As in \code{\link[=geom_point]{geom_point()}}.}

\item{width}{Maximum violin width in x slot units.}
}
\value{
A \link{Layer}.
}
\description{
Violin layer
}
