% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{plot_size}
\alias{plot_size}
\title{Set the output size}
\usage{
plot_size(width, height)
}
\arguments{
\item{width, height}{Device size in pixels.}
}
\value{
A \code{ggplot3_size} object to add to a plot with \code{+}.
}
\description{
Set the output size
}
\examples{
ggplot3(cars, aes(speed, dist)) + geom_point() + plot_size(900, 600)
}
