% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-discrete.R
\name{scale_y_discrete}
\alias{scale_y_discrete}
\title{Discrete positional scale for the y axis}
\usage{
scale_y_discrete(limits = NULL, name = NULL)
}
\arguments{
\item{limits}{Optional character vector fixing the levels and their order.}

\item{name}{Optional axis title.}
}
\value{
A \link{ScaleDiscrete} object to add with \code{+}.
}
\description{
Discrete positional scale for the y axis
}
