% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stat_count}
\alias{stat_count}
\title{Count observations at each x}
\usage{
stat_count(width = 0.8)
}
\arguments{
\item{width}{Bar width as a fraction of the x resolution.}
}
\value{
A \code{StatCount} object, to pass as a layer's \code{stat}.
}
\description{
Count observations at each x
}
