% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stat_density}
\alias{stat_density}
\title{Kernel density estimate}
\usage{
stat_density(n = 256, adjust = 1)
}
\arguments{
\item{n}{Number of evaluation points.}

\item{adjust}{Bandwidth multiplier.}
}
\value{
A \code{StatDensity} object, to pass as a layer's \code{stat}.
}
\description{
Kernel density estimate
}
