% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{stat_identity}
\alias{stat_identity}
\title{Identity statistical transformation}
\usage{
stat_identity()
}
\value{
A \code{StatIdentity} object, to pass as a layer's \code{stat}.
}
\description{
Passes layer data through untransformed. This is the default stat for
geoms that draw the data as given, such as \code{\link[=geom_point]{geom_point()}} and
\code{\link[=geom_line]{geom_line()}}.
}
