% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-special.R
\name{stat_km}
\alias{stat_km}
\title{Kaplan-Meier survival estimate}
\usage{
stat_km(conf_int = FALSE)
}
\arguments{
\item{conf_int}{Compute a 95\% confidence band (Greenwood's formula).}
}
\value{
A \code{StatKM} object, to pass as a layer's \code{stat}.
}
\description{
Kaplan-Meier survival estimate
}
