% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-special.R
\name{stat_nelson_aalen}
\alias{stat_nelson_aalen}
\title{Nelson-Aalen cumulative hazard estimate}
\usage{
stat_nelson_aalen()
}
\value{
A \code{StatNelsonAalen} object, to pass as a layer's \code{stat}.
}
\description{
Nelson-Aalen cumulative hazard estimate
}
