% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-special.R
\name{stat_waterfall}
\alias{stat_waterfall}
\title{Running totals for waterfall charts}
\usage{
stat_waterfall(width = 0.8)
}
\arguments{
\item{width}{Bar width in x slot units.}
}
\value{
A \code{StatWaterfall} object, to pass as a layer's \code{stat}.
}
\description{
Running totals for waterfall charts
}
