% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_ggnext}
\alias{theme_ggnext}
\title{The default ggnext theme}
\usage{
theme_ggnext()
}
\value{
A \link{Theme} to add to a plot with \code{+}.
}
\description{
Light background with a softly tinted panel and white gridlines.
}
\examples{
p <- ggnext(cars, aes(speed, dist)) + geom_point() + theme_dark()
}
