% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_ggplot3}
\alias{theme_ggplot3}
\title{The default ggplot3 theme}
\usage{
theme_ggplot3()
}
\value{
A \link{Theme} to add to a plot with \code{+}.
}
\description{
Light background with a softly tinted panel and white gridlines.
}
\examples{
p <- ggplot3(cars, aes(speed, dist)) + geom_point() + theme_dark()
}
