% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_minimal}
\alias{theme_minimal}
\title{Minimal theme: no panel fill, light grey gridlines}
\usage{
theme_minimal()
}
\value{
A \link{Theme} to add to a plot with \code{+}.
}
\description{
Minimal theme: no panel fill, light grey gridlines
}
