% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_void}
\alias{theme_void}
\title{Void theme: data only, no chrome at all}
\usage{
theme_void()
}
\value{
A \link{Theme} to add to a plot with \code{+}.
}
\description{
Useful for treemaps, network graphs, and sparkline-style output.
}
