% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{xlab}
\alias{xlab}
\title{Set the x axis title}
\usage{
xlab(label)
}
\arguments{
\item{label}{Axis title text.}
}
\value{
A \link{Labels} object to add to a plot with \code{+}.
}
\description{
Set the x axis title
}
