% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{xlim}
\alias{xlim}
\alias{ylim}
\title{Set continuous axis limits}
\usage{
xlim(...)

ylim(...)
}
\arguments{
\item{...}{Two numbers giving the lower and upper limit, or one length-2
numeric vector.}
}
\value{
A scale object to add to a plot with \code{+}.
}
\description{
A shorthand for \code{scale_x_continuous(limits = )}. Data outside the limits
is still computed by stats but clipped to the panel when drawn.
}
\examples{
ggplot3(cars, aes(speed, dist)) + geom_point() + xlim(0, 30) + ylim(0, 150)
}
